+++ src/modules/economy/components/DepositForm/useCreateDepositAPI.ts
@@ -1,20 +1,18 @@
 import { showNotification } from '@mantine/notifications'
-import { formatISO } from 'date-fns'
+import { DepositMethodValues } from 'modules/economy/enums'
 import { useDepositMutations } from 'modules/economy/mutations.hooks'
 import { ONGOING_DEPOSIT_INTENT_QUERY } from 'modules/economy/views'
+import { useNavigate } from 'react-router-dom'
 import { CreateDepositFormData } from './useCreateDepositLogic'
 
 export function useCreateDepositAPI(onCompletedCallback: () => void) {
   const { createDeposit } = useDepositMutations()
+  const navigate = useNavigate()
 
   async function handleSubmit(data: CreateDepositFormData) {
-    const input = {
-      ...data,
-      createdAt: formatISO(new Date()),
-    }
     return createDeposit({
       variables: {
-        input: input,
+        ...data,
       },
       refetchQueries: [ONGOING_DEPOSIT_INTENT_QUERY],
       onCompleted() {
@@ -23,7 +21,9 @@ export function useCreateDepositAPI(onCompletedCallback: () => void) {
           message: 'Innskudd har blitt opprettet',
           color: 'Green',
         })
-        onCompletedCallback()
+        if (data.depositMethod === DepositMethodValues.BANK_TRANSFER) {
+          navigate('/economy/me')
+        }
       },
       onError({ message }) {
         showNotification({
@@ -37,7 +37,7 @@ export function useCreateDepositAPI(onCompletedCallback: () => void) {
   const defaultValues = {
     amount: 0,
     description: '',
-    receipt: null,
+    depositMethod: DepositMethodValues.STRIPE,
   }
   return {
     defaultValues,