+++ src/modules/featureFlags/types.graphql.ts
@@ -1,0 +1,18 @@
+export type FeatureFlagNode = {
+  id: string
+  name: string
+  enabled: boolean
+  description: string
+}
+
+export interface AllFeatureFlagsQuery {
+  allFeatureFlags: FeatureFlagNode[]
+}
+
+export interface ToggleFeatureFlagVariables {
+  featureFlagId: string
+}
+
+export interface ToggleFeatureFlagReturns {
+  featureFlag: Pick<FeatureFlagNode, 'id'>
+}