+++ src/routes/PrivateRoutes.tsx
@@ -92,6 +92,14 @@ const AdmissionStatistics = React.lazy(
 const MySettings = React.lazy(() => import('modules/users/views/MySettings'))
 const Newbies = React.lazy(() => import('modules/users/views/Newbies'))
 
+// ==== Feature flags ====
+const FeatureFlags = React.lazy(
+  () => import('modules/featureFlags/views/FeatureFlags')
+)
+
+const EconomyDashboard = React.lazy(
+  () => import('modules/economy/views/EconomyDashboard')
+)
 export const AppRoutes: React.FC = () => {
   const { loading, error, data } = useQuery<MeQueryReturns>(ME_QUERY)
   const setUser = useStore(state => state.setUser)
@@ -423,26 +431,20 @@ export const AppRoutes: React.FC = () => {
 
         {/* ==== ECONOMY MODULE ==== */}
         <Route path="economy">
-          <Route
-            path="deposits/create"
-            element={
-              <RestrictedRoute
-                permissions={PERMISSIONS.economy.approve.deposit}
-              >
-                <CreateDeposit />
-              </RestrictedRoute>
-            }
-          />
-          <Route
-            path="deposits"
-            element={
-              <RestrictedRoute
-                permissions={PERMISSIONS.economy.approve.deposit}
-              >
-                <Deposits />
-              </RestrictedRoute>
-            }
-          />
+          <Route index element={<EconomyDashboard />} />
+          <Route path="deposits">
+            <Route
+              index
+              element={
+                <RestrictedRoute
+                  permissions={PERMISSIONS.economy.approve.deposit}
+                >
+                  <Deposits />
+                </RestrictedRoute>
+              }
+            />
+            <Route path="create" element={<CreateDeposit />} />
+          </Route>
           <Route path="me" element={<MyEconomy />} />
           <Route path="soci-products" element={<h2>Suh duh</h2>} />
 
@@ -472,6 +474,18 @@ export const AppRoutes: React.FC = () => {
           </Route>
           <Route path="*" element={<FullPage404 />} />
         </Route>
+
+        {/* ==== FEATURE FLAGS MODULE ==== */}
+        <Route
+          path="feature-flags"
+          element={
+            <RestrictedRoute
+              permissions={PERMISSIONS.featureFlags.view.featureFlag}
+            >
+              <FeatureFlags />
+            </RestrictedRoute>
+          }
+        />
 
         {/* ==== SCHEDULES MODULE ==== */}
         <Route path="schedules">