+++ src/util/featureFlagKeys.ts
@@ -1,0 +1,1 @@
+export const STRIPE_INTEGRATION_FEATURE_FLAG_KEY = 'stripe_integration'