+++ src/util/permissions.ts
@@ -28,6 +28,7 @@ export const PERMISSIONS = {
     },
     view: {
       sociSession: 'economy.view_socisession',
+      deposit: 'economy.view_deposit',
     },
     change: {
       sociSession: 'economy.change_socisession',
@@ -42,6 +43,19 @@ export const PERMISSIONS = {
       deposit: 'economy.invalidate_deposit',
     },
   },
+  // ===== FEATURE FLAGS ======
+  featureFlags: {
+    add: {
+      featureFlag: 'feature_flags.add_featureflag',
+    },
+    change: {
+      featureFlag: 'feature_flags.change_featureflag',
+    },
+    view: {
+      featureFlag: 'feature_flags.view_featureflag',
+    },
+  },
+
   // ===== HANDBOOK ======
   handbook: {
     add: {