+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # CHANGELOG
 
-## [v2023.2.1] - 2023-02-10
+## [v2023.3.1] - 2023-03-03
+
+### Changed
+
+- Economy: Deposit form bank transfer and two-step stripe deposit
+- Economy: Move navitems from sidebar to own dashboard
+
+## [v2023.2.3] - 2023-02-14
+
+### Added
+
+- Economy: Stripe deposit integration
+
+### Removed
+
+- Economy: Lists deposit view
+
+## [v2023.2.2] - 2023-02-10
 
 ### Fixed